feat: change default out to ./docs
Previously the default was output to dist/docs. This is problematic if
docs are generated before other build processes which clean the dist
folder as is the behavior with goreleaser. Goreleaser pro supports after
hooks, but that's not an answer to the default directory.

Changing the default to docs makes sense because "docs" is a commonly
used directory in the repo root for many defined project structures.
This also makes sense from a user perspective because running:

  some_command docs

one would expect a "docs" directory more than a "dist" directory
containing a "docs" subdirectory.

The caller can still override the output directory's default.