feat: options func, trim left whitespace in .md
This adds a FormatOptions function to the func map as the key `options`.
This function is implemented in markdown to trim all left whitespcae
from the options text. This allows left-alignment of the text within the
code blocks of the builtin templates, while retaining all space for user
customized templates or in the options fields of other formats like json/yaml.